fix(plugin): stop re-exporting MiniAgentPlugin from mini-agent root
Remove MiniAgentPlugin and PluginManager from mini-agent.__init__ to
break the circular import chain:
1. __init__.py imports cli.main -> calls PluginManager.discover()
2. discover() loads langfuse plugin -> from mini_agent import MiniAgentPlugin
3. MiniAgentPlugin not yet in namespace (__init__ still executing)

External plugins must now import from mini_agent.plugin subpackage:
  from mini_agent.plugin import MiniAgentPlugin

Also update mini-agent-langfuse to use the new import path.
