+++ src/mini_agent/plugin/__init__.py
@@ -1,0 +1,4 @@
+from .base import MiniAgentPlugin
+from .manager import PluginManager
+
+__all__ = ["MiniAgentPlugin", "PluginManager"]
